userspace, libc: try to use stdint types

+++ userspace/libc/src/include/sys/mouse.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include <stdint.h>
+
 struct mouse_packet {
-    unsigned long x, y, attr_byte;
+    uint32_t x, y, attr_byte;
 };
 
 #define MOUSE_ATTR_LEFT_BTN   (1 << 0)
